Set the heartbeat immediately to prevent worker dead race condition
When setting up the fetcher, a separate thread is run to initiate a
heartbeat. However, there is no guarantee that the thread will have a
chance to run before Sidekiq attempts to fetch jobs. As a result, we
saw race conditions where one thread erronenously cleaned the working
queue when it should not have.

To fix this, just set the heartbeat key initially to ensure the
heartbeat is present.

Relates to #34